+++ src/libs/nft/useNftAsset/useNftAsset.ts
@@ -14,9 +14,6 @@ function getNftCollectibleUrl(nft: any) {
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function useNftAsset(nft: any) {
-
-  console.log(nft)
-
   const metadataUrl = nft?.metadata
   const prefetchedContentType = nft?.metadata_content_type
   const image = nft?.metadata_image
@@ -48,9 +45,11 @@ export function useNftAsset(nft: any) {
   const previewSrc = previewAnimationUrl || previewImageUrl
   const properties = nftMetadata?.properties || {}
   const collectibleUrl = getNftCollectibleUrl(nft)
+  const collectionId = nft?.collectionId;
   const id = nft?.id
   const audioUrl = contentCategory === 'audio' ? RMRK1.toWeb2Url(nftMetadata?.mediaUri) : null
   const thumb = nft?.primaryResource?.thumb
+  const isRMRK2 = nft?.primaryResource ? true : false;
 
   // const stats = nft?.primaryResource?.base ? RMRK1.toWeb2Url(nft?.primaryResource?.metadata) : null
 
@@ -63,6 +62,7 @@ export function useNftAsset(nft: any) {
     previewSrc,
     name,
     id,
+    isRMRK2,
     description,
     imageUrl,
     thumb,
@@ -73,6 +73,7 @@ export function useNftAsset(nft: any) {
     contentExtension,
     contentType,
     collection,
+    collectionId,
     collectibleUrl,
     ...restMetadata,
   }
